requirements.txt
PyYAML
librosa==0.8.0
matplotlib==3.4.2
numpy
pandas==1.2.4
resampy==0.2.2
tensorboard==2.2.2
torch==1.8.1
torchaudio>=0.8.0
tqdm==4.60.0
pytorch-lightning==1.3.1
torchvision==0.9.1
